Entity Generator Refactor
