+++ AlbanianXrm.EarlyBound/Interfaces/IMyPluginFactory.cs
@@ -1,4 +1,5 @@
-﻿using AlbanianXrm.EarlyBound.Logic;
+﻿
+using AlbanianXrm.EarlyBound.Logic;
 using Syncfusion.Windows.Forms.Tools;
 
 namespace AlbanianXrm.EarlyBound.Interfaces
@@ -9,5 +10,6 @@ namespace AlbanianXrm.EarlyBound.Interfaces
         CoreToolsDownloader NewCoreToolsDownloader(MyPluginControl myPlugin);
         EntityMetadataHandler NewEntityMetadataHandler(MyPluginControl myPlugin, TreeViewAdv metadataTree);
         RelationshipMetadataHandler NewRelationshipMetadataHandler(MyPluginControl myPlugin);
+        EntityGeneratorHandler NewEntityGeneratorHandler(MyPluginControl myPlugin, TreeViewAdv metadataTree);
     }
 }
