+++ AlbanianXrm.EarlyBound/MyPluginControl.cs
@@ -28,6 +28,7 @@ namespace AlbanianXrm.EarlyBound
         private Logic.AttributeMetadataHandler AttributeMetadataHandler;
         private Logic.RelationshipMetadataHandler RelationshipMetadataHandler;
         private Logic.CoreToolsDownloader CoreToolsDownloader;
+        private Logic.EntityGeneratorHandler EntityGeneratorHandler;
 
         public MyPluginControl()
         {
@@ -35,6 +36,7 @@ namespace AlbanianXrm.EarlyBound
             MyPluginFactory = new Factories.MyPluginFactory();
             AttributeMetadataHandler = MyPluginFactory.NewAttributeMetadataHandler(this);
             CoreToolsDownloader = MyPluginFactory.NewCoreToolsDownloader(this);
+            EntityGeneratorHandler = MyPluginFactory.NewEntityGeneratorHandler(this, metadataTree);
             EntityMetadataHandler = MyPluginFactory.NewEntityMetadataHandler(this, metadataTree);
             RelationshipMetadataHandler = MyPluginFactory.NewRelationshipMetadataHandler(this);
             treeEventHandler = new TreeViewAdvBeforeCheckEventHandler(this.MetadataTree_BeforeCheck);
@@ -110,119 +112,9 @@ namespace AlbanianXrm.EarlyBound
             CoreToolsDownloader.DownloadCoreTools();
         }
 
-        private void ToolStripButton3_Click(object sender, EventArgs e)
+        private void BtnGenerateEntities_Click(object sender, EventArgs e)
         {
-
-            WorkAsync(new WorkAsyncInfo()
-            {
-                Message = $"Generating Early-Bound Classes",
-                Work = (worker, args) =>
-                {
-                    string dir = Path.GetDirectoryName(typeof(MyPluginControl).Assembly.Location).ToLower();
-                    string folder = Path.GetFileNameWithoutExtension(typeof(MyPluginControl).Assembly.Location);
-                    dir = Path.Combine(dir, folder);
-                    Process process = new Process();
-                    var connectionString = ConnectionDetail.GetConnectionStringWithPassword();
-                    process.StartInfo.Arguments = "/connectionstring:" + connectionString + (string.IsNullOrEmpty(txtNamespace.Text) ? "" : " /namespace:" + txtNamespace.Text) + " /codewriterfilter:AlbanianXrm.CrmSvcUtilExtensions.FilteringService,AlbanianXrm.CrmSvcUtilExtensions /out:" + (string.IsNullOrEmpty(txtOutputPath.Text) ? "Test.cs" : "\"" + Path.GetFullPath(txtOutputPath.Text) + "\"");
-                    process.StartInfo.WorkingDirectory = dir;
-                    process.StartInfo.UseShellExecute = false;
-                    process.StartInfo.CreateNoWindow = true;
-                    process.StartInfo.RedirectStandardOutput = true;
-                    process.StartInfo.RedirectStandardError = true;
-
-                    List<string> entities = new List<string>();
-                    List<string> allAttributes = new List<string>();
-                    List<string> allRelationships = new List<string>();
-
-                    foreach (TreeNodeAdv entity in metadataTree.Nodes)
-                    {
-                        if (entity.CheckState != CheckState.Unchecked)
-                        {
-                            EntityMetadata metadata = (EntityMetadata)entity.Tag;
-                            entities.Add(metadata.LogicalName);
-                            foreach (TreeNodeAdv item in entity.Nodes)
-                            {
-                                if (item.Text == "Attributes")
-                                {
-                                    if (item.CheckState == CheckState.Checked)
-                                    {
-                                        allAttributes.Add(metadata.LogicalName);
-                                    }
-                                    else if (item.CheckState == CheckState.Indeterminate)
-                                    {
-                                        List<string> attributes = new List<string>();
-                                        foreach (TreeNodeAdv attribute in item.Nodes)
-                                        {
-                                            if (attribute.Checked)
-                                            {
-                                                var attributeMetadata = (AttributeMetadata)attribute.Tag;
-                                                attributes.Add(attributeMetadata.LogicalName);
-                                            }
-                                        }
-                                        process.StartInfo.EnvironmentVariables.Add("AlbanianXrm.EarlyBound:Attributes:" + metadata.LogicalName, string.Join(",", attributes));
-                                    }
-                                }
-                                else if (item.Text == "Relationships")
-                                {
-                                    if (item.CheckState == CheckState.Checked)
-                                    {
-                                        allRelationships.Add(metadata.LogicalName);
-                                    }
-                                    else if (item.CheckState == CheckState.Indeterminate)
-                                    {
-                                        List<string> relationships1N = new List<string>();
-                                        List<string> relationshipsN1 = new List<string>();
-                                        List<string> relationshipsNN = new List<string>();
-                                        foreach (TreeNodeAdv relationship in item.Nodes)
-                                        {
-                                            if (relationship.Checked)
-                                            {
-                                                if (relationship.Tag is OneToManyRelationshipMetadata)
-                                                {
-                                                    var relationshipMetadata = (OneToManyRelationshipMetadata)relationship.Tag;
-                                                    if (relationshipMetadata.ReferencingEntity == metadata.LogicalName)
-                                                    {
-                                                        relationships1N.Add(relationshipMetadata.SchemaName);
-                                                    }
-                                                    else
-                                                    {
-                                                        relationshipsN1.Add(relationshipMetadata.SchemaName);
-                                                    }
-                                                }
-                                                else
-                                                {
-                                                    var relationshipMetadata = (RelationshipMetadataBase)relationship.Tag;
-                                                    relationshipsNN.Add(relationshipMetadata.SchemaName);
-                                                }
-                                            }
-                                        }
-                                        if (relationships1N.Any()) process.StartInfo.EnvironmentVariables.Add("AlbanianXrm.EarlyBound:Relationships1N:" + metadata.LogicalName, string.Join(",", relationships1N.Distinct()));
-                                        if (relationshipsN1.Any()) process.StartInfo.EnvironmentVariables.Add("AlbanianXrm.EarlyBound:RelationshipsN1:" + metadata.LogicalName, string.Join(",", relationshipsN1.Distinct()));
-                                        if (relationshipsNN.Any()) process.StartInfo.EnvironmentVariables.Add("AlbanianXrm.EarlyBound:RelationshipsNN:" + metadata.LogicalName, string.Join(",", relationshipsNN.Distinct()));
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    if (entities.Any()) process.StartInfo.EnvironmentVariables.Add("AlbanianXrm.EarlyBound:Entities", string.Join(",", entities));
-                    if (allAttributes.Any()) process.StartInfo.EnvironmentVariables.Add("AlbanianXrm.EarlyBound:AllAttributes", string.Join(",", allAttributes));
-                    if (allRelationships.Any()) process.StartInfo.EnvironmentVariables.Add("AlbanianXrm.EarlyBound:AllRelationships", string.Join(",", allRelationships));
-                    process.EnableRaisingEvents = true;
-                    process.StartInfo.FileName = Path.Combine(dir, "CrmSvcUtil.exe");
-                    process.Start();
-                    process.WaitForExit();
-                    args.Result = process.StandardOutput.ReadToEnd();
-                },
-                PostWorkCallBack = (args) =>
-                {
-                    if (args.Error != null)
-                    {
-                        MessageBox.Show(args.Error.ToString(), "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                    }
-                    MessageBox.Show((string)args.Result);
-                }
-            });
+            EntityGeneratorHandler.GenerateEntities(txtNamespace.Text, txtOutputPath.Text);
         }
 
         private void TxtOutputPath_TextChanged(object sender, EventArgs e)
